[FIX] php71: PHP 7.1 now triggers a "non well formed numeric value encountered" notice when using a string containing non-numeric chars as a number, so cast it to an int before trying to multiply it (thanks amette__)
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@61707 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/core/Tiki/MemoryLimit.php
@@ -43,7 +43,7 @@ class Tiki_MemoryLimit
 
 	private function getRaw($memory_limit)
 	{
-		$s = trim($memory_limit);
+		$s = (int) trim($memory_limit);
 		$last = strtolower($s{strlen($s)-1});
 		switch ( $last ) {
 			case 'g': $s *= 1024;
